+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.0.3] - 2019/09/21
+
+* NavigateAction tests.
+* Navigation arguments.
+
 ## [2.0.2] - 2019/09/21
 
 * UserExceptionAction.